==> 5OC/01.ColumnSelect(def).ipynb <==
import csv
import numpy as np
import pandas as pd
from pandas import DataFrame as df

with open('./Settings/Data/Default/OC_UVES_iDR5.csv') as input_csv:
    reader=csv.reader(input_csv, delimiter=',')
    head=next(reader)
    #print(head)
    autoselected=[]
    for i in range(0,len(head)):
        a=len(head[i])
        if a<=3:
            autoselected.append(i)
        else:
            pass
    #print(autoselected)
    manualselected=[0,1,14,15,20,21,43,48,49,53,54,56,57,58,59,60]
    selected=np.append(autoselected,manualselected)
    selected=np.sort(selected)
    print(selected)

data_raw=pd.read_csv('./Settings/Data/Default/OC_UVES_iDR5.csv',low_memory=False)
headers=[]
for i in range(0,len(selected)):
    headers.append(head[selected[i]])
data_select=data_raw[headers]
data_select.to_csv('./Settings/Data/data_select.csv',index=False)

err=autoselected[7:]
for i in range(0,len(err)):
    err[i]=err[i]+2
headers=[]
err.append(28)
for i in range(0,len(err)):
    headers.append(head[err[i]])
err_select=data_raw[headers]
err_select.to_csv('./Settings/Data/err_select.csv',index=False)
